+++ Localization/deDE.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-12-09 14-47-34 ]] --
+-- [[ Exported at 2022-12-13 07-11-41 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --