+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 49.4 - 2022-12-13
+### Added
+- Siege on Dragonbane Keep and Community Feast Event Reminders
+- Tooltip data for Who's a Good Bakar?
+- Experimental searching of achievement criteria; Use @ as a prefix
+
 ## 49.3 - 2022-12-09
 ### Added
 - Grand Hunts Event Reminders
 
 ### Fixed
-- Leftovers' Revenge moves from Ohn'arhran Plains Reputation to The Azure Span Reputation
+- Leftovers' Revenge moved from Ohn'arhran Plains Reputation to The Azure Span Reputation
 
 ## 49.2 - 2022-12-04
 ### Wrath Classic